refactor load
